requirements.txt
streamlit>=1.28.0
crewai>=0.28.0
pandas>=2.0.0
sqlparse>=0.4.4
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0.0
openai>=1.0.0
pandasai>=3.0.0b2
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0 